feat: move education validation schemas to a new file for better organization
