+++ backend/src/db/schema/educations.ts
@@ -57,21 +57,3 @@ export const educationsRelations = relations(educations, ({ one }) => ({
     references: [userProfile.id],
   }),
 }));
-
-// Zod schemas for validation
-export const selectEducationsSchema = createSelectSchema(educations);
-export const insertEducationsSchema = createInsertSchema(educations, {
-  schoolName: z.string().min(1, "School name is required").max(100),
-  major: z.string().min(1, "Program major is required").max(100),
-  graduated: z.boolean().default(false),
-  startDate: z.iso.datetime(),
-  endDate: z.iso.datetime().optional(),
-});
-export const updateEducationsSchema = createUpdateSchema(educations).omit({
-  userProfileId: true,
-});
-
-// Type exports
-export type Education = z.infer<typeof selectEducationsSchema>;
-export type UpdateEducation = z.infer<typeof updateEducationsSchema>;
-export type InsertEducation = z.infer<typeof insertEducationsSchema>;
